Some basic functionality

+++ src/utilities/parameters.rs
@@ -30,7 +30,7 @@ pub enum OperationalLearning {
 #[derive(Clone, Debug)]
 pub struct Parameters {
     pub number_of_repetitions: u64,
-    pub number_of_agents: u64,
+    pub number_of_agents: usize,
     pub number_of_iterations: u64,
     pub temperature_schedule: TemperatureSchedule,
     pub operational_learning: OperationalLearning,
